Display tags
This was quite a pain to get working, not least because I spent way too
long trying to build the CSS tags entirely from memory while incredibly
sleep deprived, which resulted in bollocks.

I'm very pleased with the fancy scrolling behaviour on mobile, but I'm
so unlikely to ever have that many tags it's kinda wasted. Oh well.