Sort out syntax highlighting properly
This was absolutely batshit to solve, but I've solved the problem of
code fences not generating a completely different code tree to the
{%highlight%} tag.

This also makes backtick code highlightable, which is nice.

I really don't like this as a solution because it involves parsing HTML
with regexes (HE COMES etc) but it's probably as good as I'm going to
get for now.

Could be worse - I could be on github pages! *rimshot*