Switch all links to absolute links
This is the second part of the previous commit, making sure all internal
links point to the correct canonical urls for things regardless of how
the inital page is accessed.

This should cut down on any remaining confusion between my two domains
and get rid of any pesky wwws that may have snuck in whereever.